Move RGB parsing to `rgb` color space module
The `RGBA` struct has a generic type parameter.  However, it isn't used
in the parse function, so there is nothing from which to infer the
generic type.  This makes it cumbersome to use `parse` as a class
function, so it is exposed as a module-level function instead.  This is
also consistent with the main `parse_color` function in the `parser`
module.
