+++ src/rgb.rs
@@ -1,9 +1,20 @@
 use std::fmt;
 
+use nom::{
+    branch::alt,
+    bytes::complete::{tag, tag_no_case},
+    character::complete::{char, hex_digit1, space0, space1},
+    combinator::{all_consuming, cond, opt},
+    error::ErrorKind,
+    number::complete::double,
+    Err, IResult,
+};
+
 use crate::{
     colorspace::ColorSpace,
     helper::{clamp, interpolate, mod_positive, MaxPrecision},
     hsl::HSLA,
+    parser::{legacy_alpha, legacy_separator, modern_alpha, percentage},
     types::Scalar,
     Color, Format, Fraction,
 };
@@ -229,6 +240,142 @@ impl RGBA<f64> {
     }
 }
 
+pub(crate) fn parse_rgb_color(input: &str) -> IResult<&str, Color> {
+    alt((
+        all_consuming(parse_rgb_hex),
+        all_consuming(parse_css_numeric_rgb),
+        all_consuming(parse_css_percentage_rgb),
+        all_consuming(parse_numeric_rgb),
+        all_consuming(parse_percentage_rgb),
+    ))(input.trim())
+}
+
+fn parse_rgb_hex(input: &str) -> IResult<&str, Color> {
+    fn hex_to_u8_unsafe(num: &str) -> u8 {
+        u8::from_str_radix(num, 16).unwrap()
+    }
+
+    let (input, _) = opt(char('#'))(input)?;
+    let (input, hex_chars) = hex_digit1(input)?;
+    match hex_chars.len() {
+        // RRGGBB
+        6 => {
+            let r = hex_to_u8_unsafe(&hex_chars[0..2]);
+            let g = hex_to_u8_unsafe(&hex_chars[2..4]);
+            let b = hex_to_u8_unsafe(&hex_chars[4..6]);
+            Ok((input, Color::from_rgb(r, g, b)))
+        }
+        // RGB
+        3 => {
+            let r = hex_to_u8_unsafe(&hex_chars[0..1]);
+            let g = hex_to_u8_unsafe(&hex_chars[1..2]);
+            let b = hex_to_u8_unsafe(&hex_chars[2..3]);
+            let r = r * 16 + r;
+            let g = g * 16 + g;
+            let b = b * 16 + b;
+            Ok((input, Color::from_rgb(r, g, b)))
+        }
+        // RRGGBBAA
+        8 => {
+            let r = hex_to_u8_unsafe(&hex_chars[0..2]);
+            let g = hex_to_u8_unsafe(&hex_chars[2..4]);
+            let b = hex_to_u8_unsafe(&hex_chars[4..6]);
+            let a = hex_to_u8_unsafe(&hex_chars[6..8]) as f64 / 255.0;
+            Ok((input, Color::from_rgba(r, g, b, a)))
+        }
+        // RGBA
+        4 => {
+            let r = hex_to_u8_unsafe(&hex_chars[0..1]);
+            let g = hex_to_u8_unsafe(&hex_chars[1..2]);
+            let b = hex_to_u8_unsafe(&hex_chars[2..3]);
+            let a = hex_to_u8_unsafe(&hex_chars[3..4]);
+            let r = r * 16 + r;
+            let g = g * 16 + g;
+            let b = b * 16 + b;
+            let a = (a * 16 + a) as f64 / 255.0;
+            Ok((input, Color::from_rgba(r, g, b, a)))
+        }
+        _ => Err(Err::Error(nom::error::Error::new(
+            "Expected hex string of 3 or 6 characters length",
+            ErrorKind::Many1,
+        ))),
+    }
+}
+
+fn parse_numeric_rgb(input: &str) -> IResult<&str, Color> {
+    let (input, prefixed) = opt(alt((tag("rgb("), tag("rgba("))))(input)?;
+    let is_prefixed = prefixed.is_some();
+    let (input, _) = space0(input)?;
+    let (input, r) = double(input)?;
+    let (input, _) = legacy_separator(input)?;
+    let (input, g) = double(input)?;
+    let (input, _) = legacy_separator(input)?;
+    let (input, b) = double(input)?;
+    let (input, alpha) = legacy_alpha(input)?;
+    let (input, _) = space0(input)?;
+    let (input, _) = cond(is_prefixed, char(')'))(input)?;
+
+    let r = r / 255.;
+    let g = g / 255.;
+    let b = b / 255.;
+
+    let c = Color::from_rgba_float(r, g, b, alpha);
+    Ok((input, c))
+}
+
+fn parse_css_numeric_rgb(input: &str) -> IResult<&str, Color> {
+    let (input, _) = alt((tag_no_case("rgb("), tag_no_case("rgba(")))(input)?;
+    let (input, _) = space0(input)?;
+    let (input, r) = double(input)?;
+    let (input, _) = space1(input)?;
+    let (input, g) = double(input)?;
+    let (input, _) = space1(input)?;
+    let (input, b) = double(input)?;
+    let (input, alpha) = modern_alpha(input)?;
+    let (input, _) = space0(input)?;
+    let (input, _) = char(')')(input)?;
+
+    let r = r / 255.;
+    let g = g / 255.;
+    let b = b / 255.;
+
+    let c = Color::from_rgba_float(r, g, b, alpha);
+    Ok((input, c))
+}
+
+fn parse_percentage_rgb(input: &str) -> IResult<&str, Color> {
+    let (input, prefixed) = opt(alt((tag("rgb("), tag("rgba("))))(input)?;
+    let is_prefixed = prefixed.is_some();
+    let (input, _) = space0(input)?;
+    let (input, r) = percentage(input)?;
+    let (input, _) = legacy_separator(input)?;
+    let (input, g) = percentage(input)?;
+    let (input, _) = legacy_separator(input)?;
+    let (input, b) = percentage(input)?;
+    let (input, alpha) = legacy_alpha(input)?;
+    let (input, _) = space0(input)?;
+    let (input, _) = cond(is_prefixed, char(')'))(input)?;
+
+    let c = Color::from_rgba_float(r, g, b, alpha);
+    Ok((input, c))
+}
+
+fn parse_css_percentage_rgb(input: &str) -> IResult<&str, Color> {
+    let (input, _) = alt((tag_no_case("rgb("), tag_no_case("rgba(")))(input)?;
+    let (input, _) = space0(input)?;
+    let (input, r) = percentage(input)?;
+    let (input, _) = space1(input)?;
+    let (input, g) = percentage(input)?;
+    let (input, _) = space1(input)?;
+    let (input, b) = percentage(input)?;
+    let (input, alpha) = modern_alpha(input)?;
+    let (input, _) = space0(input)?;
+    let (input, _) = char(')')(input)?;
+
+    let c = Color::from_rgba_float(r, g, b, alpha);
+    Ok((input, c))
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -322,5 +469,229 @@ mod tests {
         let c = RGBA::new(255, 127, 4);
         assert_eq!("ff7f04", c.to_hex_string(false));
         assert_eq!("#ff7f04", c.to_hex_string(true));
+    }
+
+    fn rgb(r: u8, g: u8, b: u8) -> Color {
+        Color::from_rgb(r, g, b)
+    }
+
+    fn rgba(r: u8, g: u8, b: u8, a: f64) -> Color {
+        Color::from_rgba(r, g, b, a)
+    }
+
+    fn parse_color(input: &str) -> Option<Color> {
+        parse_rgb_color(input).ok().map(|(_, c)| c)
+    }
+
+    #[test]
+    fn parse_rgb_hex_syntax() {
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("f09"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("#f09"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("#F09"));
+
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("#ff0099"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("#FF0099"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("ff0099"));
+
+        assert_eq!(Some(rgb(87, 166, 206)), parse_color("57A6CE"));
+        assert_eq!(Some(rgb(255, 0, 119)), parse_color("  #ff0077  "));
+
+        assert_eq!(None, parse_color("#1"));
+        assert_eq!(None, parse_color("#12"));
+        assert_eq!(None, parse_color("#12345"));
+        assert_eq!(None, parse_color("#1234567"));
+        assert_eq!(None, parse_color("#hh0033"));
+        assert_eq!(None, parse_color("#h03"));
+    }
+
+    #[test]
+    fn parse_rgb_hex_alpha() {
+        assert_eq!(Some(rgba(17, 51, 85, 0.6)), parse_color("#11335599"));
+        assert_eq!(Some(rgba(17, 51, 85, 0.6)), parse_color("11335599"));
+        assert_eq!(Some(rgba(17, 51, 85, 0.6)), parse_color("#1359"));
+
+        assert_eq!(Some(rgb(255, 0, 0)), parse_color("ff0000ff"));
+        assert_eq!(Some(rgb(255, 0, 0)), parse_color("#ff0000ff"));
+    }
+
+    #[test]
+    fn parse_rgb_functional_syntax() {
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(255,0,153)"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(255, 0, 153)"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb( 255 , 0 , 153 )"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(255, 0, 153.0)"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(255 0 153)"));
+
+        assert_eq!(
+            Some(rgb(255, 8, 119)),
+            parse_color("  rgb( 255  ,  8  ,  119 )  ")
+        );
+
+        assert_eq!(Some(rgb(255, 0, 127)), parse_color("rgb(100%,0%,49.8%)"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(100%,0%,60%)"));
+        assert_eq!(Some(rgb(255, 0, 119)), parse_color("rgb(100%,0%,46.7%)"));
+        assert_eq!(Some(rgb(3, 54, 119)), parse_color("rgb(1%,21.2%,46.7%)"));
+        assert_eq!(Some(rgb(255, 0, 119)), parse_color("rgb(255 0 119)"));
+        assert_eq!(
+            Some(rgb(255, 0, 119)),
+            parse_color("rgb(    255      0      119)")
+        );
+
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(100%,0%,60%)"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(100%, 0%, 60%)"));
+        assert_eq!(
+            Some(rgb(255, 0, 153)),
+            parse_color("rgb( 100% , 0% , 60% )")
+        );
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(100% 0% 60%)"));
+
+        assert_eq!(Some(rgb(100, 5, 1)), parse_color("rgb(1e2, .5e1, .5e0)"));
+        assert_eq!(Some(rgb(140, 0, 153)), parse_color("rgb(55% 0% 60%)"));
+        assert_eq!(Some(rgb(142, 0, 153)), parse_color("rgb(55.5% 0% 60%)"));
+        assert_eq!(Some(rgb(255, 0, 0)), parse_color("rgb(256,0,0)"));
+        assert_eq!(Some(rgb(255, 255, 0)), parse_color("rgb(100%,100%,-45%)"));
+
+        assert_eq!(None, parse_color("rgb(255,0)"));
+        assert_eq!(None, parse_color("rgb(255,0,0"));
+        assert_eq!(None, parse_color("rgb (256,0,0)"));
+        assert_eq!(None, parse_color("rgb(100%,0,0)"));
+        assert_eq!(None, parse_color("rgb(2550119)"));
+    }
+
+    #[test]
+    fn parse_legacy_rgb_alpha() {
+        assert_eq!(Some(rgba(10, 0, 0, 1.0)), parse_color("rgb(10,0,0,1)"));
+        assert_eq!(Some(rgba(10, 0, 0, 1.0)), parse_color("rgb(10,0,0, 1)"));
+        assert_eq!(Some(rgba(10, 0, 0, 1.0)), parse_color("rgba(10,0,0,1)"));
+        assert_eq!(Some(rgba(10, 0, 0, 1.0)), parse_color("rgba(10,0,0, 1)"));
+        assert_eq!(Some(rgba(10, 0, 0, 1.0)), parse_color("rgba(10,0,0,1.0)"));
+        assert_eq!(Some(rgba(10, 0, 0, 1.0)), parse_color("rgba(10,0,0, 1.0)"));
+
+        assert_eq!(
+            Some(rgba(10, 0, 0, 0.5)),
+            parse_color("rgba(10, 0, 0, 0.5)")
+        );
+        assert_eq!(
+            Some(rgba(10, 0, 0, 0.5)),
+            parse_color("rgba(10, 0, 0, 50%)")
+        );
+        assert_eq!(
+            Some(rgba(10, 0, 0, 0.33)),
+            parse_color("rgba(10, 0, 0, 0.33)")
+        );
+        assert_eq!(
+            Some(rgba(10, 0, 0, 0.33)),
+            parse_color("rgba(10, 0, 0, 33%)")
+        );
+    }
+
+    #[test]
+    fn parse_css_rgb_syntax() {
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(255 0 153)"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb( 255  0  153 )"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(255 0 153.0)"));
+        // `rgba` is equivalent to `rgb`
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgba(255 0 153)"));
+
+        assert_eq!(
+            Some(rgb(255, 8, 119)),
+            parse_color("  rgb( 255    8    119 )  ")
+        );
+
+        assert_eq!(Some(rgb(255, 0, 127)), parse_color("rgb(100% 0% 49.8%)"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb(100% 0% 60%)"));
+        assert_eq!(Some(rgb(255, 0, 119)), parse_color("rgb(100% 0% 46.7%)"));
+        assert_eq!(Some(rgb(3, 54, 119)), parse_color("rgb(1% 21.2% 46.7%)"));
+        assert_eq!(Some(rgb(255, 0, 119)), parse_color("rgb(255 0 119)"));
+        assert_eq!(
+            Some(rgb(255, 0, 119)),
+            parse_color("rgb(    255      0      119)")
+        );
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("rgb( 100%   0%  60% )"));
+
+        assert_eq!(Some(rgb(100, 5, 1)), parse_color("rgb(1e2 .5E1 .5e0)"));
+        assert_eq!(Some(rgb(140, 0, 153)), parse_color("rgb(55% 0% 60%)"));
+        assert_eq!(Some(rgb(142, 0, 153)), parse_color("rgb(55.5% 0% 60%)"));
+        assert_eq!(Some(rgb(255, 0, 0)), parse_color("rgb(256,0,0)"));
+        assert_eq!(Some(rgb(255, 255, 0)), parse_color("rgb(100%,100%,-45%)"));
+
+        // function names are case-insensitive
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("RGB(255 0 153)"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("RgbA(255 0 153)"));
+
+        assert_eq!(None, parse_color("rgb(255 0)"));
+        assert_eq!(None, parse_color("rgb(255 0 0"));
+        assert_eq!(None, parse_color("rgb (256 0 0)"));
+        assert_eq!(None, parse_color("rgb(100% 0 0)"));
+        assert_eq!(None, parse_color("rgb(2550119)"));
+    }
+
+    #[test]
+    fn parse_css_rgb_alpha() {
+        // alpha can be specified as a number from 0.0 to 1.0, or as a percentage
+        assert_eq!(
+            Some(rgba(10, 20, 30, 0.7)),
+            parse_color("rgb(10 20 30 / 0.7)")
+        );
+        assert_eq!(
+            Some(rgba(10, 20, 30, 0.7)),
+            parse_color("rgb(10 20 30 / 70%)")
+        );
+
+        // `rgba` is equivalent to `rgb`
+        assert_eq!(
+            Some(rgba(10, 20, 30, 0.7)),
+            parse_color("rgba(10 20 30 / 0.7)")
+        );
+
+        // spaces are not required around the '/' separator
+        assert_eq!(
+            Some(rgba(10, 20, 30, 0.7)),
+            parse_color("rgb(10 20 30/0.7)")
+        );
+        assert_eq!(
+            Some(rgba(10, 20, 30, 0.7)),
+            parse_color("rgb(10 20 30/70%)")
+        );
+
+        // extra spaces are allowed
+        assert_eq!(
+            Some(rgba(10, 20, 30, 0.7)),
+            parse_color("rgb(10   20   30  /  0.7)")
+        );
+
+        // alpha can also be combined with percentage channel values
+        assert_eq!(
+            Some(rgba(255, 0, 128, 0.7)),
+            parse_color("rgb(100% 0% 50% / 0.7)")
+        );
+
+        // an explicit 100% (or 1.0) alpha is valid
+        assert_eq!(
+            Some(rgba(10, 20, 30, 1.0)),
+            parse_color("rgb(10 20 30 / 1.0)")
+        );
+        assert_eq!(
+            Some(rgba(10, 20, 30, 1.0)),
+            parse_color("rgb(10 20 30 / 1)")
+        );
+        assert_eq!(
+            Some(rgba(10, 20, 30, 1.0)),
+            parse_color("rgb(10 20 30 / 100%)")
+        );
+    }
+
+    #[test]
+    fn parse_rgb_standalone_syntax() {
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("255,0,153"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("255, 0, 153"));
+        assert_eq!(
+            Some(rgb(255, 0, 153)),
+            parse_color("  255  ,  0  ,  153   ")
+        );
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("255 0 153"));
+        assert_eq!(Some(rgb(255, 0, 153)), parse_color("255 0 153.0"));
+
+        assert_eq!(Some(rgb(1, 2, 3)), parse_color("1,2,3"));
     }
 }
